rewrite lots of commands to ts
